move convert_service from main
